test(prd-142-wave2): W2-S4 centralize real-DB transactional fixture
Three module conftests each hardcoded the same Postgres URL (with a live
password) and duplicated a connect/begin/rollback session fixture. Lift
test_db_url / test_engine / db_session into the root orchestrator/
conftest.py (common ancestor of tests/ and modules/*/tests/), resolving
the URL through central config instead of a literal. Learning keeps its
own db_session override only to seed the Workspace FK row.
