+++ orchestrator/modules/codegraph/tests/conftest.py
@@ -9,42 +9,14 @@ import pytest
 import tempfile
 import shutil
 from pathlib import Path
-from sqlalchemy import create_engine, text
-from sqlalchemy.orm import sessionmaker, Session
+from sqlalchemy import text
+from sqlalchemy.orm import Session
 from typing import Generator
 
 from modules.codegraph import CodeGraphService
 
-
-# Test database URL (use in-memory SQLite for fast tests)
-TEST_DB_URL = "postgresql://postgres:secure_password_123@127.0.0.1:5432/orchestrator_db"
-
-
-@pytest.fixture(scope="session")
-def test_engine():
-    """Create test database engine"""
-    engine = create_engine(TEST_DB_URL)
-    yield engine
-    engine.dispose()
-
-
-@pytest.fixture(scope="function")
-def db_session(test_engine) -> Generator[Session, None, None]:
-    """
-    Provide a transactional database session for each test.
-    Rolls back after test completes.
-    """
-    connection = test_engine.connect()
-    transaction = connection.begin()
-    
-    SessionLocal = sessionmaker(bind=connection)
-    session = SessionLocal()
-    
-    yield session
-    
-    session.close()
-    transaction.rollback()
-    connection.close()
+# ``test_engine`` and the transactional ``db_session`` fixture come from the
+# root orchestrator/conftest.py (PRD-142 W2-S4) — no per-module DB URL here.
 
 
 @pytest.fixture
